Updated class

+++ API-AGRO/src/Dados/Bovinos.java
@@ -5,7 +5,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Bovinos extends Animal {
 
-	public Bovinos(double registro, Peso peso, String raca, String genero, String dataNascimento) {
+	public Bovinos(int registro, Peso peso, String raca, String genero, String dataNascimento) {
 		super(registro, peso, raca, genero, dataNascimento);
 	}
